feature: taskLogId更改为时间戳

+++ src/main/java/com/lrhealth/data/converge/dao/entity/ConvTaskLog.java
@@ -50,7 +50,7 @@ public class ConvTaskLog implements Serializable {
     /**
      * dataX或者其它引擎的任务id
      */
-    private Integer fedLogId;
+    private Long fedLogId;
 
 
 }
